agentic_native_drafting/requirements.txt
pydantic-ai>=0.2.0
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
httpx>=0.24.0
pydantic>=2.5.1
python-dotenv>=0.13.0
pytest-asyncio>=0.21.0

